Data Validation

+++ FinsharkClone/Controllers/CommentController.cs
@@ -20,6 +20,11 @@ namespace FinsharkClone.Controllers
         
         [HttpGet]
         public async Task<IActionResult> GetAll(){
+
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+
             var comments = await _commentRepository.GetAllAsync();
             var commentDtos = comments.Select(c => c.ToCommentDto());
             return Ok(commentDtos);
@@ -27,6 +32,10 @@ namespace FinsharkClone.Controllers
 
         [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id){
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepository.GetByIdAsync(id);
 
             if(comment == null){
@@ -36,8 +45,13 @@ namespace FinsharkClone.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, [FromBody] CreateCommentDto commentDto){
+
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+
             if(!await _stockRepository.StockExists(stockId)){
                 return NotFound("Stock not found");
             }
@@ -48,9 +62,13 @@ namespace FinsharkClone.Controllers
            return CreatedAtAction(nameof(GetById), new {id = commentModel.Id}, commentModel.ToCommentDto());
         }
 
-        [HttpPut("{id}")]
+        [HttpPut("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepository.UpdateAsync(id, updateDto.ToCommentFromUpdate(id));
 
             if (comment == null)
@@ -61,8 +79,12 @@ namespace FinsharkClone.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id){
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+
             var comment = await _commentRepository.DeleteAsync(id);
 
             if(comment == null){
