+++ FinsharkClone/Controllers/StockController.cs
@@ -24,7 +24,10 @@ namespace FinsharkClone.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
-        
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+
             var stocks = await _stockRepository.GetAllAsync();
             var stockDtos = stocks.Select(s => s.ToStockDto());
 
@@ -35,6 +38,10 @@ namespace FinsharkClone.Controllers
         [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+
             var stock = await _stockRepository.GetByIdAsync(id);
 
             if (stock == null)
@@ -48,6 +55,10 @@ namespace FinsharkClone.Controllers
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
         {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+
             var stockModel = stockDto.ToStockFromCreateDTO();
             await _stockRepository.CreateAsync(stockModel);
             return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto());
@@ -56,6 +67,10 @@ namespace FinsharkClone.Controllers
         [HttpPut("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockRequestDto stockDto)
         {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+
            var stockModel = await _stockRepository.UpdateAsync(id, stockDto);
 
            if (stockModel == null)
@@ -70,6 +85,10 @@ namespace FinsharkClone.Controllers
         [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+
             var stockModel = await _stockRepository.DeleteAsync(id);
 
             if (stockModel == null)
